Refactored
I split a bunch of code into seperate files to better organize it.

I changed the A11yParser so that for HTML directories it will run the file through a browser so that we can get an changes caused by JavaScript.

Also set YouTube videos to always be true for captions due to auto generation.